Clean and comment InMemoryWorldStateProxy
